backend/pyproject.toml
[project]
name = "zeropath-backend"
version = "0.1.0"
description = "ZeroPath API service – LLM-powered Python security scanner"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy>=2.0,<3",
    "alembic>=1.14,<2",
    "psycopg2-binary>=2.9,<3",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "httpx>=0.27,<1",
    "python-jose[cryptography]>=3.3,<4",
    "python-dotenv>=1.0,<2",
    "google-cloud-pubsub>=2.23,<3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9,<1",
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
